modify(minio): dev & profile에 따른 util설정
